lima/ir: print name of unsupported nir_op
lima ppir and gpir often hit unsupported alu nir_op instructions at this
stage of development. Printing the raw nir_op is not very helpful, since
the list of nir_op is an unnumbered enum that is generated during the
build, so it's annoying to figure out what is the missing
implementation.
Just print the name as pointed out by cwabbott some time ago on IRC.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>