lima: prepare pp stack registers configuration
This commit prepares the pp command stream to set the fragment shader
stack address and size registers.
These registers are part of the pp frame register.
The address set has to be different per pp, and the value is expected to
be sent to the kernel in a separate variable than the pp frame register
structure.
The size register is also set per pp, all pp get the same value as the
value in the pp frame register structure. The size register is composed
by the "stack size" and "stack offset" values.

The lima ppir still does not implement the actual use of temporaries as
of this commit, this implementation is to be done in an upcoming patch.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>