ppir: regalloc: register spilling
This implements register spilling for ppir. That is, storing some values
in memory as temporaries for when the number of available pp registers
is not enough.

The strategy for spilling is to find the most long lived register during
the program and spill it, hoping that the number of iterations spilling
registers is kept to a minimum. Then, for each use of the chosen
register, the register allocator will insert instructions to load the
value from memory before using, and in case of a write operation, insert
another instruction to store it afterwards.
To account for operations partially updating vector registers, a load
operation is also inserted before write operations.
Many of these operations might not be absolutely necessary in all cases,
but instructions are inserted anyway in order to keep the spilling
algorithm simpler.
Registers are always stored as vec4 in memory even if their use is not 4
components, again to keep the algorithm for spilling simpler.

This implementation for spilling is not the most optional possible, but
it is a working implementation. Hopefully, spilling should be left as a
last resource and effort is best spent optimizing other parts of ppir
rather than in optimizing the use of temporaries.

This code has been tested with a debug patch that forces spilling
regardless of whether the mesa register allocator succeeded or not.
It was possible to test multiple spilling iterations and it seems to
work. Also, the fragment shader shipped in the gbm-surface-render-two
gfx example requires spilling to be implemented with ppir, and it
renders successfully with this patch.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>